backup_20250726_0832/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.1",
    "google-analytics-admin>=0.24.1",
    "markdown>=3.8.2",
    "requests>=2.32.4",
]
